requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
faiss-cpu
python-dotenv
pydantic
langchain-community
faiss-cpu 
langchain-teddynote